Refactored and fixed expansion bugs
Summary:
Removed the original iterator type, and modified the find method to
return a value_type pair instead. Renamed the threadsafe iterator types
and made them friend classes. There was a bug in expansion, where all
the memory was not being modified atomically, so I put all the
table-related memory in a pointer, which is remade and atomically
swapped during expansion. Right now, the way the old pointer is deleted
is by waiting 50ms then deleting it, but there should be a better
solution. This now allows the stress test to pass with expansion
running.

Test Plan: /testing

Reviewers: binfan, dga, kaminsky

Differential Revision: https://sourcery.cmcl.cs.cmu.edu:8080/D3