Improved per-core counting
For bsd systems, there doesn't seem to be a way to get the current
cpu a thread is running on. So instead of calling sched_getcpu() to
figure out which core to increment on, we assign each thread a random
core, which it always increments. This seems to perform quite well.