Fixed configure.ac and minor optimization
The default value of CPPFLAGS was "-g -O2", but since we set it
ourselves, I removed that. Also rather than run a cuckoo_find at the
beginning of cuckoo_insert, I modified try_add_to_bucket to search the
entire bucket for the key to reduce the number of searches through the
bucket.