Upgraded to CityHash64 and other refactors
CityHash32 limits the number of buckets to 32 (and thus the number of
table items to 35), so we upped it to CityHash64. The C version was
trying to use a modified version of cityhash, which conflicted with the
new version, so I took the C version out since nothing was using it.